realized Util class with method convertsToString(Object[] a)
